Use pointer in Base32768.Decode
