+++ Base32768/Base32768.cs
@@ -1,10 +1,6 @@
-﻿#if NETSTANDARD2_1_OR_GREATER
-global using StringOrReadOnlySpanChar = System.ReadOnlySpan<char>;
-#else
-global using StringOrReadOnlySpanChar = System.String;
-#endif
-using System;
+﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Text;
 /*
@@ -40,7 +36,7 @@ namespace Kzrnm.Convert.Base32768
     /// <summary>
     /// Base32768 is a binary-to-text encoding optimised for UTF-16-encoded text.
     /// </summary>
-    public static class Base32768
+    public static partial class Base32768
     {
         internal const int BITS_PER_CHAR = 15;// Base32768 is a 15-bit encoding
         internal const int BITS_PER_BYTE = 8;
@@ -258,171 +254,5 @@ namespace Kzrnm.Convert.Base32768
 #endif
 
         #endregion Encode
-
-        #region Decode
-        /// <summary>
-        /// Decodes a Base32768.
-        /// </summary>
-        /// <param name="str">Base32768 encoded data</param>
-        /// <returns>original binary data</returns>
-        public static byte[] Decode(string str)
-#if NETSTANDARD2_1_OR_GREATER
-            => Decode(str.AsSpan());
-        /// <summary>
-        /// Decodes a Base32768.
-        /// </summary>
-        /// <param name="str">Base32768 encoded data</param>
-        /// <returns>original binary data</returns>
-        public static byte[] Decode(ReadOnlySpan<char> str)
-#endif
-        {
-            if (str.Length == 0)
-                return
-#if NETSTANDARD1_0
-                    new byte[0];
-#else
-                    Array.Empty<byte>();
-#endif
-            var res = new byte[CalculateByteLength(str)];
-            DecodeCore(str, res);
-            return res;
-        }
-
-        internal static int CalculateByteLength(StringOrReadOnlySpanChar str)
-        {
-            var length = str.Length * BITS_PER_CHAR / BITS_PER_BYTE;
-            if (str[str.Length - 1] < ZBits15Start)
-                --length;
-            return length;
-        }
-
-        internal static void DecodeCore(
-#if NETSTANDARD2_1_OR_GREATER
-            StringOrReadOnlySpanChar str, ReadOnlySpan<byte> result
-#else
-            StringOrReadOnlySpanChar str, byte[] result
-#endif
-            )
-        {
-            static void ThrowFormatException(char c)
-               => throw new FormatException($"Unrecognised Base32768 character: {c}");
-            static ushort ValidateCharAndLookupD(char c)
-            {
-                if (c < ZBits15Start || lookupD[c] is not ushort z)
-                {
-                    ThrowFormatException(c);
-                    return default;
-                }
-                return z;
-            }
-
-            var numUint8s = 0;
-            var numUint8Remaining = 8;
-            int i;
-
-            unchecked
-            {
-                for (i = 0; i + 9 < str.Length;)
-                {
-                    {
-                        var chr = str[i++];
-                        var z = ValidateCharAndLookupD(chr);
-                        result[numUint8s++] |= (byte)(z >> 7);
-                        result[numUint8s] = (byte)(z << 1);
-                    }
-                    {
-                        var chr = str[i++];
-                        var z = ValidateCharAndLookupD(chr);
-                        result[numUint8s++] |= (byte)(z >> 14);
-                        result[numUint8s++] = (byte)(z >> 6);
-                        result[numUint8s] = (byte)(z << 2);
-                    }
-                    {
-                        var chr = str[i++];
-                        var z = ValidateCharAndLookupD(chr);
-                        result[numUint8s++] |= (byte)(z >> 13);
-                        result[numUint8s++] = (byte)(z >> 5);
-                        result[numUint8s] = (byte)(z << 3);
-                    }
-                    {
-                        var chr = str[i++];
-                        var z = ValidateCharAndLookupD(chr);
-                        result[numUint8s++] |= (byte)(z >> 12);
-                        result[numUint8s++] = (byte)(z >> 4);
-                        result[numUint8s] = (byte)(z << 4);
-                    }
-                    {
-                        var chr = str[i++];
-                        var z = ValidateCharAndLookupD(chr);
-                        result[numUint8s++] |= (byte)(z >> 11);
-                        result[numUint8s++] = (byte)(z >> 3);
-                        result[numUint8s] = (byte)(z << 5);
-                    }
-                    {
-                        var chr = str[i++];
-                        var z = ValidateCharAndLookupD(chr);
-                        result[numUint8s++] |= (byte)(z >> 10);
-                        result[numUint8s++] = (byte)(z >> 2);
-                        result[numUint8s] = (byte)(z << 6);
-                    }
-                    {
-                        var chr = str[i++];
-                        var z = ValidateCharAndLookupD(chr);
-                        result[numUint8s++] |= (byte)(z >> 9);
-                        result[numUint8s++] = (byte)(z >> 1);
-                        result[numUint8s] = (byte)(z << 7);
-                    }
-                    {
-                        var chr = str[i++];
-                        var z = ValidateCharAndLookupD(chr);
-                        result[numUint8s++] |= (byte)(z >> 8);
-                        result[numUint8s++] = (byte)z;
-                    }
-                }
-
-                for (; i < str.Length; i++)
-                {
-                    var chr = str[i];
-
-                    int numZBits = 15;
-                    if (lookupD[chr] is ushort z)
-                    {
-                        if (chr < ZBits15Start)
-                        {
-                            if (i + 1 != str.Length)
-                            {
-                                ThrowFormatException(chr);
-                                return;
-                            }
-                            numZBits = 7;
-                        }
-                    }
-                    else
-                    {
-                        ThrowFormatException(chr);
-                        return;
-                    }
-
-                    do
-                    {
-                        var mask = (1 << numZBits) - 1;
-                        var zz = z & mask;
-                        if (numZBits < numUint8Remaining)
-                        {
-                            numUint8Remaining -= numZBits;
-                            result[numUint8s] |= (byte)(zz << numUint8Remaining);
-                            numZBits = 0;
-                        }
-                        else
-                        {
-                            numZBits -= numUint8Remaining;
-                            result[numUint8s++] |= (byte)(zz >> numZBits);
-                            numUint8Remaining = 8;
-                        }
-                    } while (numZBits > 0 && numUint8s < result.Length);
-                }
-            }
-        }
-#endregion Decode
     }
 }
