+++ Base32768/Base32768.csproj
@@ -9,8 +9,7 @@
 		<PackageOutputPath>$(MSBuildThisFileDirectory)..\bin\Packages\$(Configuration)\</PackageOutputPath>
 		<RootNamespace>Kzrnm.Convert.Base32768</RootNamespace>
 
-		<AllowUnsafeBlocks Condition="'$(TargetFramework)' == 'netstandard1.0'">true</AllowUnsafeBlocks>
-		<AllowUnsafeBlocks Condition="'$(TargetFramework)' == 'netstandard2.0'">true</AllowUnsafeBlocks>
+		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
 
 		<PackageId>Base32768</PackageId>
 		<Description>Base32768 is a binary encoding optimised for UTF-16-encoded text.</Description>
